Don't delete blob during rename if the new name is the same.
Closes #6347.

+++ storage/google/cloud/storage/bucket.py
@@ -926,8 +926,13 @@ class Bucket(_PropertyMixin):
         :rtype: :class:`Blob`
         :returns: The newly-renamed blob.
         """
+        same_name = blob.name == new_name
+
         new_blob = self.copy_blob(blob, self, new_name, client=client)
-        blob.delete(client=client)
+
+        if not same_name:
+            blob.delete(client=client)
+
         return new_blob
 
     @property
